Extract plugin loading logic to its own module
