+++ packages/gatsby/lib/bootstrap/load-plugins.js
@@ -1,0 +1,94 @@
+const _ = require(`lodash`)
+const slash = require(`slash`)
+const fs = require(`fs`)
+const path = require(`path`)
+
+import { store } from "../redux"
+
+module.exports = async config => {
+  // Instantiate plugins.
+  const plugins = []
+  // Create fake little site with a plugin for testing this
+  // w/ snapshots. Move plugin processing to its own module.
+  // Also test adding to redux store.
+  const processPlugin = plugin => {
+    if (_.isString(plugin)) {
+      const resolvedPath = slash(path.dirname(require.resolve(plugin)))
+      const packageJSON = JSON.parse(
+        fs.readFileSync(`${resolvedPath}/package.json`, `utf-8`)
+      )
+      return {
+        resolve: resolvedPath,
+        name: packageJSON.name,
+        version: packageJSON.version,
+        pluginOptions: {
+          plugins: [],
+        },
+      }
+    } else {
+      // Plugins can have plugins.
+      const subplugins = []
+      if (plugin.options && plugin.options.plugins) {
+        plugin.options.plugins.forEach(p => {
+          subplugins.push(processPlugin(p))
+        })
+      }
+      plugin.options.plugins = subplugins
+
+      const resolvedPath = slash(path.dirname(require.resolve(plugin.resolve)))
+      const packageJSON = JSON.parse(
+        fs.readFileSync(`${resolvedPath}/package.json`, `utf-8`)
+      )
+      return {
+        resolve: resolvedPath,
+        name: packageJSON.name,
+        version: packageJSON.version,
+        pluginOptions: _.merge({ plugins: [] }, plugin.options),
+      }
+    }
+  }
+
+  if (config.plugins) {
+    config.plugins.forEach(plugin => {
+      plugins.push(processPlugin(plugin))
+    })
+  }
+
+  // Add the site's default "plugin" i.e. gatsby-x files in root of site.
+  plugins.push({
+    resolve: slash(process.cwd()),
+    name: `defaultSitePlugin`,
+    version: `n/a`,
+    pluginOptions: {
+      plugins: [],
+    },
+  })
+
+  // Create a "flattened" array of plugins with all subplugins
+  // brought to the top-level. This simplifies running gatsby-* files
+  // for subplugins.
+  const flattenedPlugins = []
+  const extractPlugins = plugin => {
+    plugin.pluginOptions.plugins.forEach(subPlugin => {
+      flattenedPlugins.push(subPlugin)
+      extractPlugins(subPlugin)
+    })
+  }
+
+  plugins.forEach(plugin => {
+    flattenedPlugins.push(plugin)
+    extractPlugins(plugin)
+  })
+
+  store.dispatch({
+    type: `SET_SITE_PLUGINS`,
+    payload: plugins,
+  })
+
+  store.dispatch({
+    type: `SET_SITE_FLATTENED_PLUGINS`,
+    payload: flattenedPlugins,
+  })
+
+  return flattenedPlugins
+}
